+++ extensions/anytype/src/hooks/useExport.ts
@@ -1,13 +1,17 @@
 import { useCachedPromise } from "@raycast/utils";
-import { getExport } from "../api/getExport";
+import { getExport } from "../api";
+import { ExportFormat } from "../models";
 
-export function useExport(spaceId: string, objectId: string, format: string) {
+export function useExport(spaceId: string, objectId: string, format: ExportFormat) {
   const { data, error, mutate, isLoading } = useCachedPromise(
     async (spaceId, objectId, format) => {
       const response = await getExport(spaceId, objectId, format);
       return response;
     },
     [spaceId, objectId, format],
+    {
+      execute: !!spaceId && !!objectId && !!format,
+    },
   );
 
   return {