+++ extensions/sabnzbd/eslint.config.js
@@ -1,0 +1,11 @@
+const { defineConfig } = require("eslint/config");
+const raycastConfig = require("@raycast/eslint-config");
+
+module.exports = defineConfig([
+  ...raycastConfig,
+  {
+    rules: {
+      "@typescript-eslint/no-unused-vars": "off"
+    }
+  }
+]);